chore: moved User type to dir
